Refactor: consolidate duplicate code, unify types, simplify auth
- Consolidate date utils (JST_OFFSET_MS, DAY_MS, formatSectionHeader)
  into calendar-utils.ts, hoist Intl formatter to module scope
- Unify LLMProvider type (fix missing gemini), SendKeyType (cmd+enter),
  and WidgetData type into shared modules
- Merge DayEventList into DayGroup with variant prop
- Extract refreshTokens/clearAllTokens to auth/token.ts (-207 lines)
- Fix delete-account-dialog API path (/api/account -> /account)
- Remove redundant re-exports and verbose comments

16 files changed, 2 new, 1 deleted (net -419 lines)
