+++ mobile/src/auth/index.ts
@@ -1,2 +1,3 @@
 export { AuthProvider, useAuth } from "./context";
 export type { StoredUser } from "./storage";
+export { clearAllTokens, refreshTokens } from "./token";
