+++ mobile/src/components/calendar/DayGroup.tsx
@@ -1,16 +1,26 @@
-import { Text, View } from "react-native";
+import { StyleSheet, Text, View } from "react-native";
 import type { UICalendarEvent } from "../../hooks/useEvents";
 import { EventCard } from "./EventCard";
 
+/** 表示バリアント */
+type DayGroupVariant = "timeline" | "monthDetail";
+
 interface DayGroupProps {
+	/** 対象日付 */
 	date: Date;
+	/** 表示するイベントの配列 */
 	events: UICalendarEvent[];
+	/** 今日かどうか */
 	isToday: boolean;
+	/**
+	 * 表示バリアント
+	 * - "timeline": タイムライン表示（NativeWind、isTodayでアクセント強調）
+	 * - "monthDetail": 月カレンダーの日別詳細表示（背景色・ボーダー付き）
+	 * @default "timeline"
+	 */
+	variant?: DayGroupVariant;
 }
 
-/**
- * 日付のフォーマット
- */
 function formatDate(date: Date, isToday: boolean): string {
 	const weekday = date.toLocaleDateString("ja-JP", { weekday: "short" });
 	const month = date.getMonth() + 1;
@@ -22,16 +32,37 @@ function formatDate(date: Date, isToday: boolean): string {
 	return `${month}/${day}（${weekday}）`;
 }
 
-/**
- * 日別イベントグループコンポーネント
- *
- * 日付ヘッダーとその日のイベント一覧を表示します。
- * 今日の場合はアクセントカラーで強調表示されます。
- */
-export function DayGroup({ date, events, isToday }: DayGroupProps) {
+/** 日別イベントグループ（日付ヘッダー + イベント一覧） */
+export function DayGroup({
+	date,
+	events,
+	isToday,
+	variant = "timeline",
+}: DayGroupProps) {
+	const dateLabel = formatDate(date, isToday);
+
+	if (variant === "monthDetail") {
+		return (
+			<View style={monthDetailStyles.container}>
+				<View style={monthDetailStyles.header}>
+					<Text style={monthDetailStyles.headerText}>{dateLabel}</Text>
+					<Text style={monthDetailStyles.countText}>
+						{events.length > 0 ? `${events.length}件` : "予定なし"}
+					</Text>
+				</View>
+				{events.length > 0 ? (
+					events.map((event) => <EventCard key={event.id} event={event} />)
+				) : (
+					<View style={monthDetailStyles.empty}>
+						<Text style={monthDetailStyles.emptyText}>予定はありません</Text>
+					</View>
+				)}
+			</View>
+		);
+	}
+
 	return (
 		<View className="mb-4">
-			{/* 日付ヘッダー */}
 			<View
 				className={`flex-row items-center justify-between px-4 py-2 ${
 					isToday ? "mx-2 rounded-lg bg-accent-50" : ""
@@ -42,14 +73,13 @@ export function DayGroup({ date, events, isToday }: DayGroupProps) {
 						isToday ? "text-accent-600" : "text-fg"
 					}`}
 				>
-					{formatDate(date, isToday)}
+					{dateLabel}
 				</Text>
 				<Text className="text-xs text-fg-subtle">
 					{events.length > 0 ? `${events.length}件` : "予定なし"}
 				</Text>
 			</View>
 
-			{/* イベント一覧 */}
 			{events.length > 0 ? (
 				events.map((event) => <EventCard key={event.id} event={event} />)
 			) : (
@@ -62,3 +92,38 @@ export function DayGroup({ date, events, isToday }: DayGroupProps) {
 		</View>
 	);
 }
+
+/** monthDetail バリアント用スタイル */
+const monthDetailStyles = StyleSheet.create({
+	container: {
+		flex: 1,
+	},
+	header: {
+		flexDirection: "row",
+		justifyContent: "space-between",
+		alignItems: "center",
+		paddingHorizontal: 16,
+		paddingVertical: 10,
+		backgroundColor: "#FAFAFA",
+		borderBottomWidth: 1,
+		borderBottomColor: "#F5F5F5",
+	},
+	headerText: {
+		fontSize: 15,
+		fontWeight: "600",
+		color: "#404040",
+	},
+	countText: {
+		fontSize: 12,
+		color: "#A3A3A3",
+	},
+	empty: {
+		paddingHorizontal: 16,
+		paddingVertical: 24,
+	},
+	emptyText: {
+		fontSize: 13,
+		color: "#D4D4D4",
+		textAlign: "center",
+	},
+});
